Wired the back-end in and made correctionsafter the file structure reorganization
