+++ package.json
@@ -1,55 +1,57 @@
 {
-  "name": "beevr",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack-dev-server"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/benedictrobb/BEEVR.git"
-  },
-  "author": "Maja Kudlicka Antonio Trabalza",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/benedictrobb/BEEVR/issues"
-  },
-  "devDependencies": {
-    "babel-core": "^6.2.1",
-    "babel-loader": "^6.2.0",
-    "babel-preset-es2015": "^6.1.18",
-    "babel-preset-react": "^6.1.18",
-    "chai": "^3.5.0",
-    "chai-jquery": "^2.0.0",
-    "eslint": "^4.1.1",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.1.0",
-    "jsdom": "^8.1.0",
-    "mocha": "^2.4.5",
-    "react-addons-test-utils": "^0.14.7",
-    "webpack": "^1.12.9",
-    "webpack-dev-server": "^1.14.0"
-  },
-  "homepage": "https://github.com/benedictrobb/BEEVR#readme",
-  "dependencies": {
-    "babel-preset-stage-1": "^6.1.18",
-    "bcryptjs": "^2.4.3",
-    "blipp": "^2.3.0",
-    "env2": "^2.2.0",
-    "hapi": "^16.4.3",
-    "hapi-auth-cookie": "^7.0.0",
-    "inert": "^4.2.0",
-    "lodash": "^3.10.1",
-    "pg": "^6.4.0",
-    "react": "^0.14.3",
-    "react-date-range": "^0.9.4",
-    "react-dom": "^0.14.3",
-    "react-redux": "4.3.0",
-    "react-router": "^2.0.1",
-    "redux": "^3.0.4",
-    "redux-thunk": "^2.2.0"
-  }
+    "name": "beevr",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "start": "webpack-dev-server"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/benedictrobb/BEEVR.git"
+    },
+    "author": "Maja Kudlicka Antonio Trabalza",
+    "license": "ISC",
+    "bugs": {
+        "url": "https://github.com/benedictrobb/BEEVR/issues"
+    },
+    "devDependencies": {
+        "babel-core": "^6.2.1",
+        "babel-loader": "^6.2.0",
+        "babel-preset-es2015": "^6.1.18",
+        "babel-preset-react": "^6.1.18",
+        "chai": "^3.5.0",
+        "chai-jquery": "^2.0.0",
+        "eslint": "^4.1.1",
+        "eslint-plugin-import": "^2.7.0",
+        "eslint-plugin-jsx-a11y": "^6.0.2",
+        "eslint-plugin-react": "^7.1.0",
+        "jsdom": "^8.1.0",
+        "mocha": "^2.4.5",
+        "react-addons-test-utils": "^0.14.7",
+        "webpack": "^1.12.9",
+        "webpack-dev-server": "^1.14.0"
+    },
+    "homepage": "https://github.com/benedictrobb/BEEVR#readme",
+    "dependencies": {
+        "axios": "^0.16.2",
+        "babel-preset-stage-1": "^6.1.18",
+        "bcryptjs": "^2.4.3",
+        "blipp": "^2.3.0",
+        "env2": "^2.2.0",
+        "hapi": "^16.4.3",
+        "hapi-auth-cookie": "^7.0.0",
+        "inert": "^4.2.0",
+        "lodash": "^3.10.1",
+        "pg": "^6.4.0",
+        "react": "^0.14.3",
+        "react-date-range": "^0.9.4",
+        "react-dom": "^0.14.3",
+        "react-redux": "4.3.0",
+        "react-router": "^2.0.1",
+        "redux": "^3.0.4",
+        "redux-thunk": "^2.2.0"
+    },
+    "proxy": "http://trkdz-spectre-devuan:4000"
 }
