+++ src/client/actions/AppActions.js
@@ -26,8 +26,8 @@
 import bcrypt from 'bcryptjs';
 import { SET_AUTH, CHANGE_FORM, SENDING_REQUEST, SET_ERROR_MESSAGE } from '../constants/AppConstants';
 import * as errorMessages  from '../constants/MessageConstants';
-import beevrAPI from '../utils/beevrAPI.js';
-import genSalt from '../utils/salt';
+import beevrAPI from '../../utils/beevrAPI.js';
+import genSalt from '../../utils/salt';
 import { browserHistory } from 'react-router';
 
 /**
@@ -36,62 +36,62 @@ import { browserHistory } from 'react-router';
  * @param  {string} password The password of the user to be logged in
  */
 export function login(username, password) {
-  return (dispatch) => {
+    return (dispatch) => {
     // Show the loading indicator, hide the last error
-    dispatch(sendingRequest(true));
-    // If no username or password was specified, throw a field-missing error
-    if (anyElementsEmpty({ username, password })) {
-      dispatch(setErrorMessage(errorMessages.FIELD_MISSING));
-      dispatch(sendingRequest(false));
-      return;
-    }
-
-      // Use auth.js to fake a request
-      beevrAPI.login(username, hash, (success, err) => {
-        // When the request is finished, hide the loading indicator
-        dispatch(sendingRequest(false));
-        dispatch(setAuthState(success));
-        if (success === true) {
-          // If the login worked, forward the user to the dashboard and clear the form
-          forwardTo('/dashboard');
-          dispatch(changeForm({
-            username: "",
-            password: ""
-          }));
-        } else {
-          switch (err.type) {
-            case 'user-doesnt-exist':
-              dispatch(setErrorMessage(errorMessages.USER_NOT_FOUND));
-              return;
-            case 'password-wrong':
-              dispatch(setErrorMessage(errorMessages.WRONG_PASSWORD));
-              return;
-            default:
-              dispatch(setErrorMessage(errorMessages.GENERAL_ERROR));
-              return;
-          }
+        dispatch(sendingRequest(true));
+        // If no username or password was specified, throw a field-missing error
+        if (anyElementsEmpty({ username, password })) {
+            dispatch(setErrorMessage(errorMessages.FIELD_MISSING));
+            dispatch(sendingRequest(false));
+            return;
         }
-      });
+
+        // Use auth.js to fake a request
+        beevrAPI.login(username, hash, (success, err) => {
+        // When the request is finished, hide the loading indicator
+            dispatch(sendingRequest(false));
+            dispatch(setAuthState(success));
+            if (success === true) {
+                // If the login worked, forward the user to the dashboard and clear the form
+                forwardTo('/dashboard');
+                dispatch(changeForm({
+                    username: '',
+                    password: ''
+                }));
+            } else {
+                switch (err.type) {
+                case 'user-doesnt-exist':
+                    dispatch(setErrorMessage(errorMessages.USER_NOT_FOUND));
+                    return;
+                case 'password-wrong':
+                    dispatch(setErrorMessage(errorMessages.WRONG_PASSWORD));
+                    return;
+                default:
+                    dispatch(setErrorMessage(errorMessages.GENERAL_ERROR));
+                    return;
+                }
+            }
+        });
     };
-  }
+}
 
 
 /**
  * Logs the current user out
  */
 export function logout() {
-  return (dispatch) => {
-    dispatch(sendingRequest(true));
-    beevrAPI.logout((success, err) => {
-      if (success === true) {
-        dispatch(sendingRequest(false))
-        dispatch(setAuthState(false));
-        browserHistory.replace(null, '/');
-      } else {
-        dispatch(setErrorMessage(errorMessages.GENERAL_ERROR));
-      }
-    });
-  }
+    return (dispatch) => {
+        dispatch(sendingRequest(true));
+        beevrAPI.logout((success, err) => {
+            if (success === true) {
+                dispatch(sendingRequest(false));
+                dispatch(setAuthState(false));
+                browserHistory.replace(null, '/');
+            } else {
+                dispatch(setErrorMessage(errorMessages.GENERAL_ERROR));
+            }
+        });
+    };
 }
 
 /**
@@ -100,41 +100,41 @@ export function logout() {
  * @param  {string} password The password of the new user
  */
 export function register(username, password) {
-  return (dispatch) => {
+    return (dispatch) => {
     // Show the loading indicator, hide the last error
-    dispatch(sendingRequest(true));
-    // If no username or password was specified, throw a field-missing error
-    if (anyElementsEmpty({ username, password })) {
-      dispatch(setErrorMessage(errorMessages.FIELD_MISSING));
-      dispatch(sendingRequest(false));
-      return;
-    }
-
-      // Use auth.js to fake a request
-      beevrAPI.register(username, hash, (success, err) => {
-        // When the request is finished, hide the loading indicator
-        dispatch(sendingRequest(false));
-        dispatch(setAuthState(success));
-        if (success) {
-          // If the register worked, forward the user to the homepage and clear the form
-          forwardTo('/dashboard');
-          dispatch(changeForm({
-            username: "",
-            password: ""
-          }));
-        } else {
-          switch (err.type) {
-            case 'username-exists':
-              dispatch(setErrorMessage(errorMessages.USERNAME_TAKEN));
-              return;
-            default:
-              dispatch(setErrorMessage(errorMessages.GENERAL_ERROR));
-              return;
-          }
+        dispatch(sendingRequest(true));
+        // If no username or password was specified, throw a field-missing error
+        if (anyElementsEmpty({ username, password })) {
+            dispatch(setErrorMessage(errorMessages.FIELD_MISSING));
+            dispatch(sendingRequest(false));
+            return;
         }
-      });
+
+        // Use auth.js to fake a request
+        beevrAPI.register(username, hash, (success, err) => {
+        // When the request is finished, hide the loading indicator
+            dispatch(sendingRequest(false));
+            dispatch(setAuthState(success));
+            if (success) {
+                // If the register worked, forward the user to the homepage and clear the form
+                forwardTo('/dashboard');
+                dispatch(changeForm({
+                    username: '',
+                    password: ''
+                }));
+            } else {
+                switch (err.type) {
+                case 'username-exists':
+                    dispatch(setErrorMessage(errorMessages.USERNAME_TAKEN));
+                    return;
+                default:
+                    dispatch(setErrorMessage(errorMessages.GENERAL_ERROR));
+                    return;
+                }
+            }
+        });
     };
-  }
+}
 
 
 /**
@@ -142,7 +142,7 @@ export function register(username, password) {
  * @param {boolean} newState True means a user is logged in, false means no user is logged in
  */
 export function setAuthState(newState) {
-  return { type: SET_AUTH, newState };
+    return { type: SET_AUTH, newState };
 }
 
 /**
@@ -153,7 +153,7 @@ export function setAuthState(newState) {
  * @return {object}                   Formatted action for the reducer to handle
  */
 export function changeForm(newState) {
-  return { type: CHANGE_FORM, newState };
+    return { type: CHANGE_FORM, newState };
 }
 
 /**
@@ -162,7 +162,7 @@ export function changeForm(newState) {
  * @return {object}          Formatted action for the reducer to handle
  */
 export function sendingRequest(sending) {
-  return { type: SENDING_REQUEST, sending };
+    return { type: SENDING_REQUEST, sending };
 }
 
 
@@ -171,24 +171,24 @@ export function sendingRequest(sending) {
  * @param message
  */
 function setErrorMessage(message) {
-  return (dispatch) => {
-    dispatch({ type: SET_ERROR_MESSAGE, message });
+    return (dispatch) => {
+        dispatch({ type: SET_ERROR_MESSAGE, message });
 
-    const form = document.querySelector('.form-page__form-wrapper');
-    if (form) {
-      form.classList.add('js-form__err-animation');
-      // Remove the animation class after the animation is finished, so it
-      // can play again on the next error
-      setTimeout(() => {
-        form.classList.remove('js-form__err-animation');
-      }, 150);
+        const form = document.querySelector('.form-page__form-wrapper');
+        if (form) {
+            form.classList.add('js-form__err-animation');
+            // Remove the animation class after the animation is finished, so it
+            // can play again on the next error
+            setTimeout(() => {
+                form.classList.remove('js-form__err-animation');
+            }, 150);
 
-      // Remove the error message after 3 seconds
-      setTimeout(() => {
-        dispatch({ type: SET_ERROR_MESSAGE, message: '' });
-      }, 3000);
-    }
-  }
+            // Remove the error message after 3 seconds
+            setTimeout(() => {
+                dispatch({ type: SET_ERROR_MESSAGE, message: '' });
+            }, 3000);
+        }
+    };
 }
 
 /**
@@ -196,8 +196,8 @@ function setErrorMessage(message) {
  * @param {string} location The route the user should be forwarded to
  */
 function forwardTo(location) {
-  console.log('forwardTo(' + location + ')');
-  browserHistory.push(location);
+    console.log('forwardTo(' + location + ')');
+    browserHistory.push(location);
 }
 
 
@@ -207,10 +207,10 @@ function forwardTo(location) {
  * @return {boolean}         True if there are empty elements, false if there aren't
  */
 function anyElementsEmpty(elements) {
-  for (let element in elements) {
-    if (!elements[element]) {
-      return true;
+    for (let element in elements) {
+        if (!elements[element]) {
+            return true;
+        }
     }
-  }
-  return false;
+    return false;
 }
