+++ src/client/components/pages/Dashboard.js
@@ -1,38 +1,44 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
+import Axios from 'axios'
 
-console.log('Dashboard');
+//console.log('Dashboard');
 class Dashboard extends Component {
     constructor() {
         super();
-        this.state = { items: [] }
+        this.state = { 
+            data: {} 
+        };
     }
 
     componentWillMount() {
-        fetch('http://trkdz-spectre-devuan:4000')
-            .then( response => response.json() ) 
-            .then( ({results: items}) => this.setState({items}))
+        Axios.get('/api')
+            .then((response) => {
+                console.log(response.data);
+                var data = response.data;
+                this.setState({data});
+            });
     }
-     
+        
     render() {
-        let items = this.state.items
+        let data = this.state.data;
         return (
             <article>
                 <section className="text-section">
                     <p>Student</p>
                     <p>Resident</p>
+                    <h1>{data.message}</h1>
                 </section>
-            {items.map(item => <h1>{item.message}</h1>)}
             </article>
         );
-  }
+    }
 }
 
 // Which props do we want to inject, given the global state?
 function select(state) {
-  return {
-    data: state
-  };
+    return {
+        data: state
+    };
 }
 
 // Wrap the component to inject dispatch and state into it
