+++ src/client/components/pages/LoginPage.js
@@ -9,22 +9,22 @@
 import React, { Component} from 'react';
 import { connect } from 'react-redux';
 import Form from '../Form.js';
-import beevrAPI from '../../utils/beevrAPI.js';
+import beevrAPI from '../../../utils/beevrAPI.js';
 import { login } from '../../actions/AppActions';
 import Loadingindicator from '../Loadingindicator.js';
 
 class LoginPage extends Component {
 
-    componentWillMount() {
+    /*componentWillMount() {
     fetch( 'https://swapi.co/api/people/?format=json' )
       .then( response => response.json() )
       .then( ({results: items}) => this.setState({items}))
-    }
+    }*/
 
     render() {
         const dispatch = this.props.dispatch;
         const { formState, currentlySending } = this.props.data;
-    return (
+        return (
             <div className="form-page__wrapper">
                 <div className="form-page__form-wrapper">
                     <div className="form-page__form-header">
@@ -32,12 +32,12 @@ class LoginPage extends Component {
                     </div>
                     {/* While the form is sending, show the loading indicator,
                         otherwise show "Log in" on the submit button */}
-                <Form data={formState} dispatch={dispatch} location={location} history={this.props.history}
-          onSubmit={this._login.bind(this)} btnText={"Login"} currentlySending={currentlySending}/>
+                    <Form data={formState} dispatch={dispatch} location={location} history={this.props.history}
+                        onSubmit={this._login.bind(this)} btnText={'Login'} currentlySending={currentlySending}/>
                 </div>
             </div>
         );
-  }
+    }
 
 
     _login(username, password) {
@@ -47,9 +47,9 @@ class LoginPage extends Component {
 
 // Which props do we want to inject, given the global state?
 function select(state) {
-  return {
-    data: state
-  };
+    return {
+        data: state
+    };
 }
 
 // Wrap the component to inject dispatch and state into it
