+++ src/client/reducers/index.js
@@ -12,43 +12,43 @@
 
 import { CHANGE_FORM, SET_AUTH, SENDING_REQUEST, SET_ERROR_MESSAGE } from '../constants/AppConstants';
 
-const assign = Object.assign
-import beevrAPI from '../utils/beevrAPI.js';
+const assign = Object.assign;
+import beevrAPI from '../../utils/beevrAPI.js';
 
 
 const initialState = {
-  formState: {
-    username: '',
-    password: ''
-  },
-  currentlySending: false,
-  loggedIn: beevrAPI.loggedIn(),
-  errorMessage: ''
+    formState: {
+        username: '',
+        password: ''
+    },
+    currentlySending: false,
+    loggedIn: beevrAPI.loggedIn(),
+    errorMessage: ''
 };
 
 // Takes care of changing the application state
 export default function homeReducer(state = initialState, action) {
-  switch (action.type) {
+    switch (action.type) {
     case CHANGE_FORM:
-      return assign({}, state, {
-        formState: action.newState
-      });
-      break;
+        return assign({}, state, {
+            formState: action.newState
+        });
+        break;
     case SET_AUTH:
-      return assign({}, state, {
-        loggedIn: action.newState
-      });
-      break;
+        return assign({}, state, {
+            loggedIn: action.newState
+        });
+        break;
     case SENDING_REQUEST:
-      return assign({}, state, {
-        currentlySending: action.sending
-      });
-      break;
+        return assign({}, state, {
+            currentlySending: action.sending
+        });
+        break;
     case SET_ERROR_MESSAGE:
-      return assign({}, state, {
-        errorMessage: action.message
-      });
+        return assign({}, state, {
+            errorMessage: action.message
+        });
     default:
-      return state;
-  }
+        return state;
+    }
 }
